+++ spec/flame/sentry_context/version_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-describe 'Flame::RavenContext::VERSION' do
+describe 'Flame::SentryContext::VERSION' do
 	subject { Object.const_get(self.class.description) }
 
 	it { is_expected.to match(/^\d+\.\d+\.\d+$/) }