+++ spec/spec_helper.rb
@@ -3,11 +3,12 @@
 require 'pry-byebug'
 
 require 'simplecov'
-SimpleCov.start
 
-if ENV['CODECOV_TOKEN']
-	require 'codecov'
-	SimpleCov.formatter = SimpleCov::Formatter::Codecov
+if ENV['CI']
+	require 'simplecov-cobertura'
+	SimpleCov.formatter = SimpleCov::Formatter::CoberturaFormatter
 end
 
-require_relative '../lib/flame/raven_context'
+SimpleCov.start
+
+require_relative '../lib/flame/sentry_context'